+++ yandex/resource_yandex_mdb_mongodb_cluster.go
@@ -341,6 +341,22 @@ func prepareCreateMongodbRequest(d *schema.ResourceData, meta *Config) (*mongodb
 	}
 
 	switch ver := cfgVer.(string); ver {
+	case "5.0":
+		{
+			configSpec.MongodbSpec = &mongodb.ConfigSpec_MongodbSpec_5_0{
+				MongodbSpec_5_0: &mongodb.MongodbSpec5_0{
+					Mongod: &mongodb.MongodbSpec5_0_Mongod{
+						Resources: &res,
+					},
+					Mongos: &mongodb.MongodbSpec5_0_Mongos{
+						Resources: &res,
+					},
+					Mongocfg: &mongodb.MongodbSpec5_0_MongoCfg{
+						Resources: &res,
+					},
+				},
+			}
+		}
 	case "4.4":
 		{
 			configSpec.MongodbSpec = &mongodb.ConfigSpec_MongodbSpec_4_4{
@@ -734,6 +750,24 @@ func extractMongodbResources(version string, mongo *mongodb.ClusterConfig) ([]ma
 	}
 
 	switch version {
+	case "5.0":
+		{
+			mongocfg := mongo.Mongodb.(*mongodb.ClusterConfig_Mongodb_5_0).Mongodb_5_0
+			d := mongocfg.Mongod
+			if d != nil {
+				return flattenMongoDBResources(d.Resources)
+			}
+
+			s := mongocfg.Mongos
+			if s != nil {
+				return flattenMongoDBResources(s.Resources)
+			}
+
+			cfg := mongocfg.Mongocfg
+			if cfg != nil {
+				return flattenMongoDBResources(cfg.Resources)
+			}
+		}
 	case "4.4":
 		{
 			mongocfg := mongo.Mongodb.(*mongodb.ClusterConfig_Mongodb_4_4).Mongodb_4_4
@@ -862,6 +896,22 @@ func getMongoDBClusterUpdateRequest(d *schema.ResourceData) (*mongodb.UpdateClus
 	securityGroupIds := expandSecurityGroupIds(d.Get("security_group_ids"))
 
 	switch d.Get("cluster_config.0.version").(string) {
+	case "5.0":
+		{
+			req := &mongodb.UpdateClusterRequest{
+				ClusterId:   d.Id(),
+				Description: d.Get("description").(string),
+				Labels:      labels,
+				ConfigSpec: &mongodb.ConfigSpec{
+					Version:           d.Get("cluster_config.0.version").(string),
+					MongodbSpec:       expandMongoDBSpec5_0(d),
+					BackupWindowStart: expandMongoDBBackupWindowStart(d),
+					Access:            &mongodb.Access{DataLens: d.Get("cluster_config.0.access.0.data_lens").(bool)},
+				},
+				SecurityGroupIds: securityGroupIds,
+			}
+			return req, nil
+		}
 	case "4.4":
 		{
 			req := &mongodb.UpdateClusterRequest{
@@ -928,7 +978,7 @@ func getMongoDBClusterUpdateRequest(d *schema.ResourceData) (*mongodb.UpdateClus
 		}
 	default:
 		{
-			return nil, fmt.Errorf("wrong MongoDB version: required either 4.4, 4.2, 4.0 or 3.6, got %s", d.Get("cluster_config.version"))
+			return nil, fmt.Errorf("wrong MongoDB version: required either 5.0, 4.4, 4.2, 4.0 or 3.6, got %s", d.Get("cluster_config.version"))
 		}
 	}
 }
@@ -963,6 +1013,10 @@ func updateMongodbClusterParams(d *schema.ResourceData, meta interface{}) error 
 
 	if d.HasChange("resources") {
 		switch d.Get("cluster_config.0.version").(string) {
+		case "5.0":
+			{
+				updatePath = append(updatePath, "config_spec.mongodb_spec_5_0")
+			}
 		case "4.4":
 			{
 				updatePath = append(updatePath, "config_spec.mongodb_spec_4_4")
@@ -981,7 +1035,7 @@ func updateMongodbClusterParams(d *schema.ResourceData, meta interface{}) error 
 			}
 		default:
 			{
-				return fmt.Errorf("wrong MongoDB version: required either 4.4, 4.2, 4.0 or 3.6, got %s", d.Get("cluster_config.version"))
+				return fmt.Errorf("wrong MongoDB version: required either 5.0, 4.4, 4.2, 4.0 or 3.6, got %s", d.Get("cluster_config.version"))
 			}
 		}